requirements.txt
pandas
matplotlib
scikit-learn
numpy
seaborn
django
tensorflow
opencv-python
google-generativeai
djangorestframework
python-dotenv
googlemaps
django-allauth
celery 
redis
